Switching to a different coordinate system to support maps with size > 8 and sets
Eas able to improve the coordinate system and seems to be working for
every case now.
The coords are the same as before, a vector of positional indexes from the root but for maps instead of an index it will be for :

    map key : a string K followed by the hash of the key form
    map value: a string V followed by the hash of the key form for the val

For sets it will also be a string K followed by the hash of the set
element form.

As an example :

(defn foo [a b]
  (assoc {1 10
          (+ 42 43) 100}
         :x #{(+ 1 2) (+ 3 4) (+ 4 5) (+ 1 1 (* 2 2))}))

some examples coordinates :

    [3 1 "K-240379483"]   => (+ 42 43)
    [3 2 "K1305480196" 3] => (* 2 2)

I tested with a couple of applications and it looks fine.

+++ src/jvm/clojure/storm/Emitter.java
@@ -216,7 +216,7 @@ public class Emitter {
 
 		StringBuilder strCoordBuilder = new StringBuilder();
 		for (int i = 0; i < coord.count(); i++) {
-			strCoordBuilder.append((int)coord.nth(i));
+			strCoordBuilder.append(coord.nth(i));
 			if(i < (coord.count()-1))
 				strCoordBuilder.append(",");
 		}
